pyproject.toml
[tool.poetry]
name = "neo4j-llm-course"
version = "0.1.0"
description = ""
authors = ["Quentin Reul <Quentin.H.Reul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
langchain = "^0.1"
langchainhub = "^0.1"
langchain-openai = "^0.1"
openai = "^1.14"
python = "^3.11"
python-dotenv = "^1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
